adding typescript support and updating the version of reactjs and nextjs
