+++ package.json
@@ -1,33 +1,36 @@
 {
-  "name": "breeze-next",
-  "version": "0.1.0",
-  "private": true,
-  "scripts": {
-    "dev": "next dev",
-    "build": "next build",
-    "start": "next start"
-  },
-  "dependencies": {
-    "@headlessui/react": "^1.4.2",
-    "axios": "^0.21.1",
-    "next": "^12.0.4",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "swr": "^1.2.2"
-  },
-  "devDependencies": {
-    "@babel/eslint-parser": "^7.12.1",
-    "@next/eslint-plugin-next": "^10.0.4",
-    "@tailwindcss/forms": "^0.5.2",
-    "autoprefixer": "^10.4.2",
-    "babel-eslint": "^10.1.0",
-    "eslint": "^7.17.0",
-    "eslint-config-prettier": "^7.1.0",
-    "eslint-plugin-next": "^0.0.0",
-    "eslint-plugin-prettier": "^3.3.0",
-    "eslint-plugin-react": "^7.22.0",
-    "postcss": "^8.4.6",
-    "prettier": "2.2.1",
-    "tailwindcss": "^3.0.22"
-  }
+    "name": "breeze-next",
+    "version": "0.1.0",
+    "private": true,
+    "scripts": {
+        "dev": "next dev",
+        "build": "next build",
+        "start": "next start"
+    },
+    "dependencies": {
+        "@headlessui/react": "^1.4.2",
+        "axios": "^0.21.1",
+        "next": "12.2.5",
+        "react": "18.2.0",
+        "react-dom": "18.2.0",
+        "swr": "^1.2.2"
+    },
+    "devDependencies": {
+        "@babel/eslint-parser": "^7.12.1",
+        "@next/eslint-plugin-next": "^10.0.4",
+        "@tailwindcss/forms": "^0.5.2",
+        "@types/node": "^18.7.14",
+        "@types/react": "^18.0.18",
+        "autoprefixer": "^10.4.2",
+        "babel-eslint": "^10.1.0",
+        "eslint": "^7.17.0",
+        "eslint-config-prettier": "^7.1.0",
+        "eslint-plugin-next": "^0.0.0",
+        "eslint-plugin-prettier": "^3.3.0",
+        "eslint-plugin-react": "^7.22.0",
+        "postcss": "^8.4.6",
+        "prettier": "2.2.1",
+        "tailwindcss": "^3.0.22",
+        "typescript": "^4.8.2"
+    }
 }
