+++ src/components/ApplicationLogo.tsx
