+++ src/components/AuthCard.tsx
