--- src/components/Label.js
@@ -1,9 +1,0 @@
-const Label = ({ className, children, ...props }) => (
-    <label
-        className={`${className} block font-medium text-sm text-gray-700`}
-        {...props}>
-        {children}
-    </label>
-)
-
-export default Label
