+++ src/components/NavLink.tsx
@@ -1,6 +1,11 @@
 import Link from 'next/link'
+import { LinkProps } from 'next/dist/client/link'
+import { ReactNode } from 'react'
 
-const NavLink = ({ active = false, children, ...props }) => (
+const NavLink = ({ active = false, children, ...props }: {
+    active: boolean,
+    children: ReactNode
+} & LinkProps) => (
     <Link {...props}>
         <a
             className={`inline-flex items-center px-1 pt-1 border-b-2 text-sm font-medium leading-5 focus:outline-none transition duration-150 ease-in-out ${
