+++ src/components/ResponsiveNavLink.tsx
@@ -1,6 +1,11 @@
 import Link from 'next/link'
+import { LinkProps } from 'next/dist/client/link'
+import { ReactNode } from 'react'
 
-const ResponsiveNavLink = ({ active = false, children, ...props }) => (
+const ResponsiveNavLink = ({ active = false, children, ...props }: {
+    active: boolean,
+    children: ReactNode
+} & LinkProps) => (
     <Link {...props}>
         <a
             className={`block pl-3 pr-4 py-2 border-l-4 text-base font-medium leading-5 focus:outline-none transition duration-150 ease-in-out ${
@@ -15,7 +20,7 @@ const ResponsiveNavLink = ({ active = false, children, ...props }) => (
 
 export const ResponsiveNavButton = props => (
     <button
-        className="block w-full pl-3 pr-4 py-2 border-l-4 text-left text-base font-medium leading-5 focus:outline-none transition duration-150 ease-in-out border-transparent text-gray-600 hover:text-gray-800 hover:bg-gray-50 hover:border-gray-300 focus:text-gray-800 focus:bg-gray-50 focus:border-gray-300"
+        className='block w-full pl-3 pr-4 py-2 border-l-4 text-left text-base font-medium leading-5 focus:outline-none transition duration-150 ease-in-out border-transparent text-gray-600 hover:text-gray-800 hover:bg-gray-50 hover:border-gray-300 focus:text-gray-800 focus:bg-gray-50 focus:border-gray-300'
         {...props}
     />
 )
