+++ src/hooks/auth.tsx
@@ -3,7 +3,10 @@ import axios from '@/lib/axios'
 import { useEffect } from 'react'
 import { useRouter } from 'next/router'
 
-export const useAuth = ({ middleware, redirectIfAuthenticated } = {}) => {
+export const useAuth = ({
+    middleware,
+    redirectIfAuthenticated,
+}: { middleware?: string; redirectIfAuthenticated?: string } = {}) => {
     const router = useRouter()
 
     const { data: user, error, mutate } = useSWR('/api/user', () =>
